+++ SwiftCompilerSources/Sources/CMakeLists.txt
@@ -6,6 +6,9 @@
 # See http://swift.org/LICENSE.txt for license information
 # See http://swift.org/CONTRIBUTORS.txt for Swift project authors
 
+# NOTE: Subdirectories must be added in dependency order.
+
+add_subdirectory(Basic)
 if(SWIFT_ENABLE_EXPERIMENTAL_STRING_PROCESSING)
   add_subdirectory(ExperimentalRegex)
 endif()
