+++ SwiftCompilerSources/Sources/SIL/Function.swift
@@ -10,6 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
+import Basic
 import SILBridging
 
 final public class Function : CustomStringConvertible, HasName {
