+++ SwiftCompilerSources/Sources/SIL/Instruction.swift
@@ -10,6 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
+import Basic
 import SILBridging
 
 //===----------------------------------------------------------------------===//
