+++ SwiftCompilerSources/Sources/SIL/Value.swift
@@ -10,6 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
+import Basic
 import SILBridging
 
 public protocol Value : AnyObject, CustomStringConvertible {
