+++ include/swift/Basic/BridgedSwiftObject.h
@@ -15,8 +15,8 @@
 //
 //===----------------------------------------------------------------------===//
 
-#ifndef SWIFT_SIL_BRIDGEDSWIFTOBJECT_H
-#define SWIFT_SIL_BRIDGEDSWIFTOBJECT_H
+#ifndef SWIFT_BASIC_BRIDGEDSWIFTOBJECT_H
+#define SWIFT_BASIC_BRIDGEDSWIFTOBJECT_H
 
 #if defined(__OpenBSD__)
 #include <sys/stdint.h>
@@ -34,20 +34,20 @@
 #if __has_feature(nullability)
 // Provide macros to temporarily suppress warning about the use of
 // _Nullable and _Nonnull.
-# define SWIFT_BEGIN_NULLABILITY_ANNOTATIONS                                   \
+#define SWIFT_BEGIN_NULLABILITY_ANNOTATIONS                                   \
   _Pragma("clang diagnostic push")                                             \
   _Pragma("clang diagnostic ignored \"-Wnullability-extension\"")
-# define SWIFT_END_NULLABILITY_ANNOTATIONS                                     \
+#define SWIFT_END_NULLABILITY_ANNOTATIONS                                     \
   _Pragma("clang diagnostic pop")
 
 #else
 // #define _Nullable and _Nonnull to nothing if we're not being built
 // with a compiler that supports them.
-# define _Nullable
-# define _Nonnull
-# define _Null_unspecified
-# define SWIFT_BEGIN_NULLABILITY_ANNOTATIONS
-# define SWIFT_END_NULLABILITY_ANNOTATIONS
+#define _Nullable
+#define _Nonnull
+#define _Null_unspecified
+#define SWIFT_BEGIN_NULLABILITY_ANNOTATIONS
+#define SWIFT_END_NULLABILITY_ANNOTATIONS
 #endif
 
 SWIFT_BEGIN_NULLABILITY_ANNOTATIONS
