+++ include/swift/SIL/SILBridgingUtils.h
@@ -22,30 +22,6 @@
 
 namespace swift {
 
-inline BridgedStringRef getBridgedStringRef(llvm::StringRef str) {
-  return { (const unsigned char *)str.data(), str.size() };
-}
-
-inline StringRef getStringRef(BridgedStringRef str) {
-  return StringRef((const char *)str.data, str.length);
-}
-
-/// Copies the string in an malloc'ed memory and the caller is responsible for
-/// freeing it.
-inline BridgedStringRef getCopiedBridgedStringRef(std::string str,
-                                           bool removeTrailingNewline = false) {
-  // A couple of mallocs are needed for passing a std::string to Swift. But
-  // it's currently only used or debug descriptions. So, its' maybe not so bad -
-  // for now.
-  // TODO: find a better way to pass std::strings to Swift.
-  StringRef strRef(str);
-  if (removeTrailingNewline)
-    strRef.consume_back("\n");
-  llvm::MallocAllocator allocator;
-  StringRef copy = strRef.copy(allocator);
-  return getBridgedStringRef(copy);
-}
-
 inline SILLocation getSILLocation(BridgedLocation loc) {
   return reinterpret_cast<SILDebugLocation *>(&loc)->getLocation();
 }
