+++ include/swift/module.modulemap
@@ -1,3 +1,9 @@
+module BasicBridging {
+  header "Basic/BridgedSwiftObject.h"
+  header "Basic/BasicBridging.h"
+  export *
+}
+
 module SILBridging {
   header "SIL/SILBridging.h"
   export *
