+++ lib/SIL/Utils/SILBridging.cpp
@@ -10,6 +10,7 @@
 //
 //===----------------------------------------------------------------------===//
 
+#include "swift/Basic/BridgingUtils.h"
 #include "swift/SIL/SILNode.h"
 #include "swift/SIL/SILBridgingUtils.h"
 #include "swift/SIL/SILGlobalVariable.h"
@@ -137,19 +138,6 @@ void registerBridgedClass(BridgedStringRef className, SwiftMetatype metatype) {
     abort();
   }
   nodeMetatypes[(unsigned)kind] = metatype;
-}
-
-//===----------------------------------------------------------------------===//
-//                            Bridging C functions
-//===----------------------------------------------------------------------===//
-
-void OStream_write(BridgedOStream os, BridgedStringRef str) {
-  static_cast<raw_ostream *>(os.streamAddr)->write((const char*)(str.data), str.length);
-}
-
-/// Frees a string which was allocated by getCopiedBridgedStringRef.
-void freeBridgedStringRef(BridgedStringRef str) {
-  llvm::MallocAllocator().Deallocate(str.data, str.length);
 }
 
 //===----------------------------------------------------------------------===//
